+++ src/Swarm/Game/Scenario.hs
@@ -44,25 +44,18 @@ module Swarm.Game.Scenario (
 
   -- * Loading from disk
   loadScenario,
-  ScenarioCollection (..),
-  scenarioCollectionToList,
-  ScenarioItem (..),
-  _SISingle,
-  scenarioItemName,
-  loadScenarios,
+  loadScenarioFile,
 ) where
 
 import Control.Algebra (Has)
 import Control.Arrow ((&&&))
 import Control.Carrier.Lift (Lift, sendIO)
-import Control.Carrier.Throw.Either (Throw, runThrow, throwError)
+import Control.Carrier.Throw.Either (Throw, throwError)
 import Control.Lens hiding (from, (<.>))
-import Control.Monad (filterM, unless, when)
+import Control.Monad (filterM, when)
 import Data.Aeson.Key qualified as Key
 import Data.Aeson.KeyMap (KeyMap)
 import Data.Aeson.KeyMap qualified as KeyMap
-import Data.Char (isSpace)
-import Data.List ((\\))
 import Data.Map (Map)
 import Data.Map qualified as M
 import Data.Maybe (listToMaybe)
@@ -73,7 +66,7 @@ import Data.Yaml as Y
 import GHC.Generics (Generic)
 import GHC.Int (Int64)
 import Linear.V2
-import Paths_swarm (getDataDir, getDataFileName)
+import Paths_swarm (getDataFileName)
 import Swarm.Game.Entity
 import Swarm.Game.Recipe
 import Swarm.Game.Robot (TRobot, trobotName)
@@ -81,8 +74,8 @@ import Swarm.Game.Terrain
 import Swarm.Language.Pipeline (ProcessedTerm)
 import Swarm.Util (reflow)
 import Swarm.Util.Yaml
-import System.Directory (doesDirectoryExist, doesFileExist, listDirectory)
-import System.FilePath (takeBaseName, takeExtensions, (<.>), (</>))
+import System.Directory (doesFileExist)
+import System.FilePath ((<.>), (</>))
 import Witch (from, into)
 
 ------------------------------------------------------------
@@ -332,99 +325,6 @@ loadScenario scenario em = do
     Nothing -> throwError @Text $ "Scenario not found: " <> from @String scenario
     Just fileName -> loadScenarioFile em fileName
 
--- | A scenario item is either a specific scenario, or a collection of
---   scenarios (*e.g.* the scenarios contained in a subdirectory).
-data ScenarioItem = SISingle Scenario | SICollection Text ScenarioCollection
-
--- | Retrieve the name of a scenario item.
-scenarioItemName :: ScenarioItem -> Text
-scenarioItemName (SISingle s) = s ^. scenarioName
-scenarioItemName (SICollection name _) = name
-
--- | A scenario collection is a tree of scenarios, keyed by name,
---   together with an optional order.  Invariant: every item in the
---   scOrder exists as a key in the scMap.
-data ScenarioCollection = SC
-  { scOrder :: Maybe [FilePath]
-  , scMap :: Map FilePath ScenarioItem
-  }
-
--- | Convert a scenario collection to a list of scenario items.
-scenarioCollectionToList :: ScenarioCollection -> [ScenarioItem]
-scenarioCollectionToList (SC Nothing m) = M.elems m
-scenarioCollectionToList (SC (Just order) m) = (m M.!) <$> order
-
--- | Load all the scenarios from the scenarios data directory.
-loadScenarios :: (Has (Lift IO) sig m) => EntityMap -> m (Either Text ScenarioCollection)
-loadScenarios em = runThrow $ do
-  dataDir <- sendIO getDataDir
-  loadScenarioDir em (dataDir </> "scenarios")
-
-orderFileName :: FilePath
-orderFileName = "00-ORDER.txt"
-
--- | Recursively load all scenarios from a particular directory, and also load
---   the 00-ORDER file (if any) giving the order for the scenarios.
-loadScenarioDir ::
-  (Has (Lift IO) sig m, Has (Throw Text) sig m) =>
-  EntityMap ->
-  FilePath ->
-  m ScenarioCollection
-loadScenarioDir em dir = do
-  let orderFile = dir </> orderFileName
-      dirName = takeBaseName dir
-  orderExists <- sendIO $ doesFileExist orderFile
-  morder <- case orderExists of
-    False -> do
-      when (dirName /= "Testing") $
-        sendIO . putStrLn $
-          "Warning: no " <> orderFileName <> " file found in " <> dirName
-            <> ", using alphabetical order"
-      return Nothing
-    True -> Just . filter (not . null) . lines <$> sendIO (readFile orderFile)
-  fs <- sendIO $ keepYamlOrDirectory <$> listDirectory dir
-
-  case morder of
-    Just order -> do
-      let missing = fs \\ order
-          dangling = order \\ fs
-
-      unless (null missing) $
-        sendIO . putStr . unlines $
-          ( "Warning: while processing " <> (dirName </> orderFileName) <> ": files not listed in "
-              <> orderFileName
-              <> " will be ignored"
-          ) :
-          map ("  - " <>) missing
-
-      unless (null dangling) $
-        sendIO . putStr . unlines $
-          ( "Warning: while processing " <> (dirName </> orderFileName)
-              <> ": nonexistent files will be ignored"
-          ) :
-          map ("  - " <>) dangling
-    Nothing -> pure ()
-
-  -- Only keep the files from 00-ORDER.txt that actually exist.
-  let morder' = filter (`elem` fs) <$> morder
-  SC morder' . M.fromList <$> mapM (\item -> (item,) <$> loadScenarioItem em (dir </> item)) fs
- where
-  keepYamlOrDirectory = filter (\f -> takeExtensions f `elem` ["", ".yaml"])
-
--- | Load a scenario item (either a scenario, or a subdirectory
---   containing a collection of scenarios) from a particular path.
-loadScenarioItem ::
-  (Has (Lift IO) sig m, Has (Throw Text) sig m) =>
-  EntityMap ->
-  FilePath ->
-  m ScenarioItem
-loadScenarioItem em path = do
-  isDir <- sendIO $ doesDirectoryExist path
-  let collectionName = into @Text . dropWhile isSpace . takeBaseName $ path
-  case isDir of
-    True -> SICollection collectionName <$> loadScenarioDir em path
-    False -> SISingle <$> loadScenarioFile em path
-
 -- | Load a scenario from a file.  The @Maybe Seed@ argument is a
 --   seed provided by the user (either on the command line, or
 --   specified through the UI), if any.
@@ -438,9 +338,3 @@ loadScenarioFile em fileName = do
   case res of
     Left parseExn -> throwError @Text (from @String (prettyPrintParseException parseExn))
     Right c -> return c
-
-------------------------------------------------------------
--- Some lenses + prisms
-------------------------------------------------------------
-
-makePrisms ''ScenarioItem
